Move streaming reducer to scnn and add WSS reducer heads
